Added index, about pages, included matter.js and style.css
Initialised a matter.js engine and renderer and added a few objects: a ball that can jump and roll around with WASD, and a rectangle that acts like a button on collision from the ball: causing a white opening to form on the canvas. This opening is not yet traversable, but the next commit or the one after that should have it lead to another environment in about.html (which is currently empty).