Update index.html
Placed the door and player in the same negative-value collisionFilter group so that they cannot collide, and applied transparency to the door so that the player would still be visible from underneath it.

Matter.js doesn't seem to have any native way to determine which objects sit at the front, so the transparency is a compromise. In the next commit, the player will be made darker as it enters the door for greater visual contrast.